ai_data_readiness/api/requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# GraphQL dependencies
strawberry-graphql[fastapi]==0.215.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Async support
asyncpg==0.29.0
aiofiles==23.2.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# System monitoring
psutil==5.9.6

# Data processing
pandas==2.1.4
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
click==8.1.7